fix: propagate swarm id and share docker client
