+++ moderator-service/Dockerfile
@@ -4,6 +4,7 @@ WORKDIR /app
 # copy parent pom and module sources
 COPY pom.xml ./
 COPY common/topology-core/pom.xml common/topology-core/pom.xml
+COPY common/docker-client/pom.xml common/docker-client/pom.xml
 COPY test/pom.xml test/pom.xml
 COPY generator-service/pom.xml generator-service/pom.xml
 COPY moderator-service/pom.xml moderator-service/pom.xml
@@ -16,6 +17,7 @@ COPY orchestrator-service/pom.xml orchestrator-service/pom.xml
 COPY swarm-controller-service/pom.xml swarm-controller-service/pom.xml
 COPY scenario-manager-service/pom.xml scenario-manager-service/pom.xml
 COPY common/topology-core/src common/topology-core/src
+COPY common/docker-client/src common/docker-client/src
 COPY test/src test/src
 COPY generator-service/src generator-service/src
 COPY moderator-service/src moderator-service/src
