+++ orchestrator-service/pom.xml
@@ -61,6 +61,11 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <groupId>io.pockethive</groupId>
+      <artifactId>docker-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>net.logstash.logback</groupId>
       <artifactId>logstash-logback-encoder</artifactId>
       <version>${logstash.logback.version}</version>
