+++ orchestrator-service/src/main/java/io/pockethive/orchestrator/infra/docker/DockerConfiguration.java
@@ -5,6 +5,7 @@ import com.github.dockerjava.core.DefaultDockerClientConfig;
 import com.github.dockerjava.core.DockerClientImpl;
 import com.github.dockerjava.transport.DockerHttpClient;
 import com.github.dockerjava.httpclient5.ApacheDockerHttpClient;
+import io.pockethive.docker.DockerContainerClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
